Feat(#48): axios객체 생성, 습관체크 메ㅑ
