+++ package.json
@@ -13,6 +13,7 @@
     "@types/node": "^16.18.101",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
+    "axios": "^1.7.3",
     "chart.js": "^4.4.3",
     "chartjs-plugin-datalabels": "^2.2.0",
     "dayjs": "^1.11.12",
@@ -20,6 +21,7 @@
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.3.1",
     "react-emotion": "^10.0.0",
+    "react-query": "^3.39.3",
     "react-router-dom": "^6.25.1",
     "react-scripts": "5.0.1",
     "react-styled": "^0.0.5",
@@ -58,5 +60,6 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  }
+  },
+  "proxy": "http://3.39.60.18:9000"
 }
