+++ src/pages/Home/Home.tsx
@@ -7,7 +7,6 @@ import TopBar from "../../components/organs/Bars/TopBar";
 
 import { useNavigate } from "react-router-dom";
 
-import home from "../../apis/home";
 
 const Home = () => {
   const navigate = useNavigate();
